devex: shuffle code around to provide 100% coverage
